start work on better seed/task management

+++ src/seeding.rs
@@ -5,7 +5,7 @@ use rand::SeedableRng;
 /// Generate a vector of `nseeds` unique [`u64`] values.
 /// These values can be used as unique seeds for different replicates
 /// of a simulation.
-pub fn make_unique_seeds(initial_seed: u64, nseeds: usize) -> Vec<u64> {
+pub fn make_unique_seeds(initial_seed: u64, nseeds: i32) -> Vec<u64> {
     let rseed = rand_distr::Uniform::new(0_u64, u64::MAX);
     let mut rng = StdRng::seed_from_u64(initial_seed);
     let mut rv = vec![];
